Alertme/requirements.txt
PyQt6==6.5.0
Flask==2.3.2
Flask-Cors==4.0.0
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.31.0
twilio==8.2.0
firebase-admin==6.1.0
google-auth==2.19.1
openai==0.27.8
PyJWT==2.7.0
Flask-JWT-Extended==4.5.2
